fixed the switch case; fixed the description; organized the code
